made changes according to last PR review
