+++ src/components/TeamCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import styles from '../../assets/css/groupsView.module.scss';
-import Arrow from '../../assets/images/groups/teamCardArrow.svg';
+import styles from '../assets/css/groupsView.module.scss';
+import Arrow from '../assets/images/groups/teamCardArrow.svg';
 import { Link } from 'react-router-dom';
 export default function TeamCard({
   name,
@@ -14,7 +14,7 @@ export default function TeamCard({
       <div className={styles.TeamName}>
         <h1>{name}</h1>
         <button>
-          <Link to={{ pathname: `/groups/${slug}`, state: id }}>
+          <Link to={`/groups/${slug}`}>
             <img src={Arrow} alt=""></img>
           </Link>
         </button>
